int_tech2_cw2/requirements.txt
Flask==2.0.3
Werkzeug==2.0.3
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.31
geocoder==1.38.1
requests==2.26.0
python-dotenv==0.19.1
